+++ R/open_curly_linter.R
@@ -29,8 +29,15 @@ open_curly_linter <- function(allow_single_line = FALSE) {
              some_after <- length(tokens_after) %!=% 0L
 
              content_after <- unname(substr(line, parsed$col1 + 1L, nchar(line)))
+             content_before <- unname(substr(line, 1, parsed$col1 - 1L))
 
              only_comment <- rex::re_matches(content_after, rex::rex(any_spaces, "#", something, end))
+
+             double_curly <- rex::re_matches(content_after, rex::rex(start, "{")) || rex::re_matches(content_before, rex::rex("{", end))
+
+             if (double_curly) {
+               return()
+             }
 
              whitespace_after <-
                unname(substr(line, parsed$col1 + 1L, parsed$col1 + 1L)) %!=% ""